% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionnal_cdf.R
\name{build_u_tree}
\alias{build_u_tree}
\title{Conditional cumulative density function}
\usage{
build_u_tree(u, longitudinalDVine)
}
\arguments{
\item{u}{(numeric) a vector of marginal uniforms}

\item{longitudinalDVine}{A S3 object of class LongitudinalDVine}
}
\value{
(numeric) The triangle of conditionnal uniforms
}
\description{
Computing the conditionnal cdf of an observation, knowing the
surrounding ones.
}
\examples{
x <- system.file("extdata", "eg1.dat", package = "DVineSD")
x <- data.table::fread(x, skip = 22, select = "LOAD")$LOAD
trained_DVine <- fit_LongitudinalDVine(head(x, 100), max_lag = 20L)
u <- ecdf(x)(x[1:5])
build_u_tree(u, trained_DVine)
}
