% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_LongitudinalDVine.R
\name{fit_LongitudinalDVine}
\alias{fit_LongitudinalDVine}
\title{Fitting a D-Vine copula on a longitudinal data set.}
\usage{
fit_LongitudinalDVine(x, pmargin, ..., max_lag)
}
\arguments{
\item{x}{(numeric) a vector of longitudinal observations}

\item{pmargin}{(function) the cdf of the marginal distribution}

\item{...}{arguments of pmargin}

\item{max_lag}{(integer) the maximum lag the dependence structure can support}
}
\value{
an S3 object containing a list of all fitted bivariate copulas.
As an attribute, you can retrieve a version of the \code{pmargin} value with
its parameters nested inside the function.
}
\description{
Fitting a D-Vine copula on a longitudinal data set.
}
\examples{
# Simulating an AR(1) time series
x <- stats::arima.sim(list(order = c(1,0,0), ar = 0.85), n = 100)
x <- pnorm(x)
trained_DVine <- fit_LongitudinalDVine(x)

# Fit a longitudinal Dvine on 
x <- system.file("extdata/eg1.dat", package = "DVineSD")
x <- data.table::fread(x, skip = 22, select = "LOAD")$LOAD

}
