% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalDVine_class.R
\docType{class}
\name{LongitudinalDVine_class}
\alias{LongitudinalDVine_class}
\alias{LongitudinalDVine_dist}
\title{Class to create an instance of a longitudinal D-Vine copula structure.}
\usage{
LongitudinalDVine_dist(x)
}
\arguments{
\item{x}{(list) an embedded list describing the D-vine structure.
Each element of the list corresponds to the main characteristics of the
\code{bicop_dist()} function from the \code{rvinecopulib} package, which are :
\itemize{
\item \code{family} (characters),
\item \code{rotation} (An integer between 0 and 365),
\item \code{parameters} (A named vector).
}

See Examples for more details.}
}
\value{
An S4 object that contains the D-vine copula structure.
}
\description{
Class to create an instance of a longitudinal D-Vine copula structure.
}
\examples{
LongitudinalDVine_dist(x = list(
   list(
      family = "gaussian",
      rotation = 0,
      parameters = 0.85
   ),
   list(
      family = "indep",
      rotation = 0,
      parameters = numeric(0)
   )))

}
