% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul_serial_dependence.R
\name{simul_serial_dependence}
\alias{simul_serial_dependence}
\title{Simulate realisations of a time series}
\usage{
simul_serial_dependence(
  longitudinalDVine,
  x,
  qmargin,
  nsim = 1L,
  nseq = 100L,
  seed = 32L,
  cores = 4L,
  ...
)
}
\arguments{
\item{longitudinalDVine}{A S3 object of class LongitudinalDVine}

\item{x}{(numeric) a vector of marginal uniforms on which we condition}

\item{qmargin}{(function) the marginal quantile function of x}

\item{nsim}{(integer) the desired number of simulations}

\item{nseq}{(integer) the length of the simulated sequence}

\item{seed}{(integer) a simulation seed}

\item{cores}{(integer) the number of available threads to do parallel computation}

\item{...}{further arguments to qmargin}
}
\value{
(numeric) a matrix of dimension nsim x n_seq of simulated values
}
\description{
Simulating time-dependant sequence
}
\examples{
Dvine_copula <- DVineSD::LongitudinalDVine_dist(x = list(
   list(
      family = "clayton",
      rotation = 0,
      parameters = 0.9
   ),
   list(
      family = "indep",
      rotation = 0,
      parameters = numeric(0)
   )
))
x <- DVineSD::simul_serial_dependence(Dvine_copula, nseq = 10L)
}
